+++ src/frontend/pages/api/shipping.ts
@@ -1,7 +1,7 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
 import InstrumentationMiddleware from '../../utils/telemetry/InstrumentationMiddleware';
 import ShippingGateway from '../../gateways/rpc/Shipping.gateway';
-import { CartItem, Empty, Money } from '../../protos/demo';
+import { Address, CartItem, Empty, Money } from '../../protos/demo';
 import CurrencyGateway from '../../gateways/rpc/Currency.gateway';
 
 type TResponse = Money | Empty;
@@ -9,8 +9,9 @@ type TResponse = Money | Empty;
 const handler = async ({ method, query }: NextApiRequest, res: NextApiResponse<TResponse>) => {
   switch (method) {
     case 'GET': {
-      const { itemList = '', currencyCode = 'USD' } = query;
-      const { costUsd } = await ShippingGateway.getShippingCost(JSON.parse(itemList as string) as CartItem[]);
+      const { itemList = '', currencyCode = 'USD', address = '' } = query;
+      const { costUsd } = await ShippingGateway.getShippingCost(JSON.parse(itemList as string) as CartItem[],
+          JSON.parse(address as string) as Address);
       const cost = await CurrencyGateway.convert(costUsd!, currencyCode as string);
 
       return res.status(200).json(cost!);